+++ packages/chaingraph-types/src/node/node-enums.ts
@@ -15,10 +15,6 @@ export enum NodeStatus {
   Ready = 'ready',
   Pending = 'pending',
   Executing = 'executing',
-  /**
-   * Node is currently executing one or mode background actions.
-   */
-  Backgrounding = 'backgrounding',
   Completed = 'completed',
   Skipped = 'skipped',
   Error = 'error',